requirements.txt
torch>=2.8.0
torchvision
torchaudio
torchdiffeq
lightning>=2.5.3

# JAX ecosystem
jax>=0.6.2
jaxlib>=0.6.2
flax>=0.10.7
optax>=0.2.5

# --- Model Components ---
einops
safetensors
transformers
wandb

# --- Bioinformatics ---
scanpy>=1.11.4
anndata>=0.11.4
squidpy>=1.6.5
scvi-tools>=1.3.3
tangram-sc>=1.0.4
python-igraph>=1.0.0
leidenalg>=0.11.0
ray>=2.53.0
novosparc>=0.4.4

# --- Data & Visualization ---
numpy>=2.2.6
pandas>=2.3.2
scipy>=1.15.3
scikit-learn>=1.7.1
matplotlib>=3.10.5
seaborn>=0.13.2
networkx>=3.4.2
opencv-python
pillow
timm

# --- Utils ---
tqdm
pyyaml
h5py